Restructure "Single responsibility" example
The example for the 'Single responsibility' in 'OOP Principles' lesson is written in a contradictory way which may confuse students:

- The example firstly states the incorrect way by saying "So instead of this";

- Then proceeds to show the correct way of solving the issue by saying "You should extract all the DOM manipulation into its own module _and use it like so_";

- Lastly, contradicts the above by saying "In fact - the function isGameOver shouldn’t be calling the DOM function anyway. That should go elsewhere (directly in the game-loop).";

The example must make clear that there are two issues with the given code block and address each problem separately to avoid the apparent contradiction. By doing so, I suggest adding a new code block for a `Game` module, inside which the `game-loop`function is contains the respective conditional statement to check `gameOver`.